fix: converted a bunch of components to hooks they were initially components because for hooks they needed to be down the hierarchy
