--- packages/ps2/src/components/Navigation/UpdateChecker/constants.ts
@@ -1,1 +1,0 @@
-export const VERSION_CHECK_INTERVAL = 10 * 60_000;
